move base test class into haystack core
